+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fixed #41: Renamed `WsjcppCore::startsWith` to `wsjcpp::starts_with`
 - Renamed `WsjcppCore::endsWith` to `wsjcpp::ends_with`
 - Fixed #39: Removed missing implementation `WsjcppCore::listOfDirs`
+- Renamed `WsjcppCore::getListOfDirs` to `wsjcpp::directory_list`
 
 ## [v0.2.5] - 2026-05-19 (2026 May 19)